addng quasi temporary solution to expose service
